Update patch set 1
Patch Set 1:

Thanks for the link.
However, I don't think this is an option, as the author of the patch clearly states it isn't fully confirmed yet. The entire Exynos mobicore/trustzone driver depends on curl, so I'm afraid the patch you linked might break it easily.
I'm no expert on the curl/ssl relations either and I guess it could be hard finding someone who is.

Patch-set: 1